requirements.txt
discord.py==2.4.0
aiohttp==3.11.11
python-dotenv==1.0.1
flask==3.1.0
